+++ R/java.R
@@ -6,6 +6,14 @@ ch.api <- function(){
   J("org.corehunter.API")
 }
 
+# --------- #
+# CONSTANTS #
+# --------- #
+
+ch.constants <- function(){
+  J("org.corehunter.util.CoreHunterConstants")
+}
+
 # ---- #
 # DATA #
 # ---- #
@@ -17,7 +25,7 @@ ch.distances <- function(){
   J("org.corehunter.data.DistanceMatrixData")
 }
 ch.genotypes <- function(){
-  J("org.corehunter.data.GenotypeData")
+  J("org.corehunter.data.FrequencyGenotypeData")
 }
 ch.phenotypes <- function(){
   J("uno.informatics.data.dataset.FeatureData")