+++ NEWS.md
@@ -1,6 +1,16 @@
 Core Hunter 3
 =============
 
+Version 3.2.0 (08/09/2017)
+--------------------------
+
+ - New option to specify a set of always and/or never selected accessions.
+ - Added support for step-based in addition to time-based stop conditions.
+ - Using `set.seed` prior to executing Core Hunter, in combination with step-based stop conditions, now yields reproducible results.
+ - Largely reduced memory footprint of genotype and distance data.
+ - Default maximum time without improvement is now only applied in case no explicit stop conditions have been specified.
+ - Fixed issue when loading phenotype data with a single trait.
+
 Version 3.1.0 (27/01/2017)
 --------------------------